Add hunger effect logic
Introduces hunger effect application based on nuScore in HungerService, with various potion effects for different score ranges. Adds player cache management for nuScores, updates listeners to handle cache on join/quit, and refactors EatListener to apply effects after eating.

+++ src/main/kotlin/de/schnitzel/nutrition/util/getNuScore.kt
@@ -3,6 +3,6 @@ package de.schnitzel.nutrition.util
 fun getNuScore(nutritionData: NutritionData): Int {
     val nuScore =
         (nutritionData.meat + nutritionData.dairy + nutritionData.sugar + nutritionData.cereals + nutritionData.fruit) / 5
-    return nuScore.toInt()
+    return nuScore
 
 }
